Замени библиотеку gson на jackson при работе с json файлами. Объединены классы работы с файлами. Добавлен logger lombok. Улучшена читаемость кода.

+++ src/test/java/MainClassTests.java
@@ -6,30 +6,38 @@ import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.PrintStream;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class MainClassTests {
-    private final String inputFilePath = "src/test/resources/in/players.xml";
+    private final String inputFilePath = "src/test/resources/players.xml";
     private final String outputFilePath = "src/test/resources/result/players.json";
     @Test
     void MainWithCorrectArgsTest() {
-        val args = new String[]{inputFilePath, outputFilePath};
-        Main.main(args);
-        val outputFile = new File(outputFilePath);
-        assertTrue(outputFile.exists());
-        outputFile.delete();
+        String[] args = {inputFilePath, outputFilePath};
+
+        try {
+            Main.main(args);
+
+            val outputFile = new File(outputFilePath);
+            assertTrue(outputFile.exists());
+        } finally {
+            val outputFile = new File(outputFilePath);
+            if (outputFile.exists()) {
+                outputFile.delete();
+            }
+        }
     }
 
     @Test
     void MainWithIncorrectArgsTest() {
         val errContent = new ByteArrayOutputStream();
         System.setErr(new PrintStream(errContent));
+
         try {
-            val args = new String[]{inputFilePath};
+            String[] args = {inputFilePath};
             Main.main(args);
 
-            assertThat(errContent.toString()).contains("Получено некорректное количество аргументов");
+            assertTrue(errContent.toString().contains("Получено некорректное количество аргументов"));
         } finally {
             System.setErr(System.err);
         }
